#15 fix: 인스턴스 주입으로 수정
