+++ UmbrellaFriend-iOS/UmbrellaFriend-iOS/Presentation/Umbrella/UmbrellaRentScene/View/UmbrellaRentBottomSheetView.swift
@@ -11,19 +11,15 @@ import SnapKit
 
 final class UmbrellaRentBottomSheetView: UIView {
     
-    // MARK: - Properties
-    
-    private var bottomHeight: CGFloat = SizeLiterals.Screen.screenHeight * 501 / 812
-    
     // MARK: - UI Components
     
-    private let backgroundView: UIView = {
+    let backgroundView: UIView = {
         let view = UIView()
         view.backgroundColor = .umbrellaBlack.withAlphaComponent(0.6)
         return view
     }()
     
-    private let bottomSheetView: UIView = {
+    let bottomSheetView: UIView = {
         let view = UIView()
         view.backgroundColor = .umbrellaWhite
         view.layer.maskedCorners = [.layerMinXMinYCorner, .layerMaxXMinYCorner]
@@ -117,13 +113,6 @@ final class UmbrellaRentBottomSheetView: UIView {
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
-    
-    override func layoutSubviews() {
-        super.layoutSubviews()
-        
-        showBottomSheet()
-        setDismissAction()
-    }
 }
 
 // MARK: - Extensions
@@ -202,46 +191,6 @@ private extension UmbrellaRentBottomSheetView {
             $0.width.equalTo((SizeLiterals.Screen.screenWidth - 40) / 2)
             $0.height.equalTo(SizeLiterals.Screen.screenHeight * 54 / 812)
         }
-    }
-    
-    func showBottomSheet() {
-        DispatchQueue.main.async {
-            self.bottomSheetView.snp.remakeConstraints {
-                $0.leading.trailing.bottom.equalToSuperview()
-                $0.top.equalToSuperview().inset(SizeLiterals.Screen.screenHeight - self.bottomHeight)
-            }
-            UIView.animate(withDuration: 0.3, delay: 0, options: .curveEaseInOut, animations: {
-                self.backgroundView.backgroundColor = .umbrellaBlack.withAlphaComponent(0.6)
-                self.layoutIfNeeded()
-            })
-        }
-    }
-    
-    func hideBottomSheet() {
-        DispatchQueue.main.async {
-            self.bottomSheetView.snp.remakeConstraints {
-                $0.leading.trailing.bottom.equalToSuperview()
-            }
-            UIView.animate(withDuration: 0.3, delay: 0, options: .curveEaseInOut, animations: {
-                self.backgroundView.backgroundColor = .clear
-                self.layoutIfNeeded()
-            }, completion: { _ in
-            })
-        }
-    }
-    
-    func setDismissAction() {
-        let tapGesture = UITapGestureRecognizer(target: self, action: #selector(hideBottomSheetAction))
-        backgroundView.addGestureRecognizer(tapGesture)
-        
-        let swipeGesture = UISwipeGestureRecognizer(target: self, action: #selector(hideBottomSheetAction))
-        swipeGesture.direction = .down
-        self.addGestureRecognizer(swipeGesture)
-    }
-    
-    @objc
-    func hideBottomSheetAction() {
-        hideBottomSheet()
     }
 }
